+++ src/main/java/net/oneandone/lavender/modules/NodeResource.java
@@ -55,8 +55,8 @@ public class NodeResource extends Resource {
         return path;
     }
 
-    public long getLastModified() {
-        return lastModified;
+    public String getContentId() {
+        return Long.toString(lastModified);
     }
 
     public boolean isOutdated() {