+++ src/main/java/net/oneandone/lavender/modules/Resource.java
@@ -24,7 +24,8 @@ import java.security.NoSuchAlgorithmException;
 
 public abstract class Resource {
     public abstract String getPath();
-    public abstract long getLastModified() throws IOException;
+
+    public abstract String getContentId() throws IOException;
 
     /** for logging purpose */
     public abstract String getOrigin();